refactored and split sell-offer folder from buy-offer folder
