+++ src/subcommand/wallet/buy_offer.rs
@@ -4,14 +4,14 @@ pub mod accept;
 pub mod create;
 
 #[derive(Debug, Parser)]
-pub(crate) enum Offer {
+pub(crate) enum BuyOffer {
   #[command(about = "Accept offer to buy inscription")]
   Accept(accept::Accept),
   #[command(about = "Create offer to buy inscription")]
   Create(create::Create),
 }
 
-impl Offer {
+impl BuyOffer {
   pub(crate) fn run(self, wallet: Wallet) -> SubcommandResult {
     match self {
       Self::Accept(accept) => accept.run(wallet),
