+++ tests/wallet.rs
@@ -5,6 +5,7 @@ mod authentication;
 mod balance;
 mod batch_command;
 mod burn;
+mod buy_offer;
 mod cardinals;
 mod create;
 mod dump;
@@ -12,7 +13,6 @@ mod inscribe;
 mod inscriptions;
 mod label;
 mod mint;
-mod offer;
 mod outputs;
 #[cfg(unix)]
 mod pending;
@@ -23,6 +23,7 @@ mod resume;
 mod runics;
 mod sats;
 mod selection;
+mod sell_offer;
 mod send;
 mod sign;
 mod split;
